+++ hosts/mellon/default.nix
@@ -34,8 +34,6 @@
 
   device.type = "desktop";
 
-  networking.networkmanager.enable = true;
-
   i18n.defaultLocale = "en_US.UTF-8";
   console.keyMap = "us";
   time.timeZone = "America/Sao_Paulo";