+++ hosts/mellon/maintenance.nix
@@ -12,5 +12,27 @@
     };
 
     power-profiles-daemon.enable = true;
+
+    snapper = {
+      snapshotInterval = "hourly";
+      cleanupInterval = "1d";
+      persistentTimer = true;
+
+      configs.root = {
+        SUBVOLUME = "/";
+        TIMELINE_CREATE = true;
+        TIMELINE_CLEANUP = true;
+        TIMELINE_LIMIT_HOURLY = 24;
+        TIMELINE_LIMIT_DAILY = 7;
+        TIMELINE_LIMIT_WEEKLY = 4;
+        TIMELINE_LIMIT_MONTHLY = 12;
+        TIMELINE_LIMIT_YEARLY = 2;
+      };
+    };
+
+    journald.extraConfig = ''
+      SystemMaxUse=2G
+      MaxRetentionSec=1month
+    '';
   };
 }